Fix: Simplex.propagate_eq bug
In [propagate_eq x y], when [x] is not a solved dependent variable, it
might still appear in the solution of [y]. Therefore, it is not
correct to implement the propagation using [R.fuse x (R.find y)],
since it would lead to an ill-formed solution substitution where [x]
appears in its solution.

Instead, in this case fall back to adding a general equality.